zcash_client_backend: move FindAccountForAddressError to data_api::error
Relocate the error type alongside the other wallet error types. Update
call sites in zcash_client_sqlite and zcash_client_memory accordingly.
