+++ zcash_client_backend/src/data_api/error.rs
@@ -440,3 +440,43 @@ impl<DE, TE, SE, FE, CE, N> From<pczt::roles::tx_extractor::Error>
         Error::Pczt(PcztError::Extraction(e))
     }
 }
+
+/// Errors that may occur when resolving the account controlling an address.
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum FindAccountForAddressError<E> {
+    /// Error returned by the underlying wallet backend.
+    Backend(E),
+
+    /// A Unified Address whose receivers map to different accounts.
+    UnifiedAddressConflict,
+}
+
+impl<E> From<E> for FindAccountForAddressError<E> {
+    fn from(err: E) -> Self {
+        Self::Backend(err)
+    }
+}
+
+impl<E: Display> Display for FindAccountForAddressError<E> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            FindAccountForAddressError::Backend(e) => {
+                write!(f, "Wallet backend error: {e}")
+            }
+            FindAccountForAddressError::UnifiedAddressConflict => write!(
+                f,
+                "Receivers of the provided Unified Address map to different wallet accounts."
+            ),
+        }
+    }
+}
+
+impl<E: error::Error + 'static> error::Error for FindAccountForAddressError<E> {
+    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+        match self {
+            FindAccountForAddressError::Backend(e) => Some(e),
+            FindAccountForAddressError::UnifiedAddressConflict => None,
+        }
+    }
+}
